Make return type match up to end of class name
Fixes the syntax highlighting part of #23.

I'll be honest in saying that I'm just starting out with Swift so my knowledge of syntax isn't the greatest but this was really bothering me. If there's some case where there's multiple words after the `->` this isn't valid but if the syntax is only ever `-> ClassName {` this should word.